+++ .eslintrc.js
@@ -3,7 +3,7 @@ module.exports = {
   extends: ['@webpack-contrib/eslint-config-webpack', 'prettier'],
   overrides: [
     {
-      files: ['lib/**/*.js'],
+      files: ['src/**/*.js'],
       parserOptions: {
         sourceType: 'script',
       },
