+++ .gitignore
@@ -15,3 +15,4 @@ Thumbs.db
 .vscode
 *.sublime-project
 *.sublime-workspace
+lib
