+++ .prettierignore
@@ -3,3 +3,4 @@
 /node_modules
 /test/fixtures
 CHANGELOG.md
+lib
